[Canvas][tech-debt] Update Redux components to reflect new structure (#73844)
