+++ x-pack/plugins/canvas/public/components/page_manager/page_manager.component.tsx
