+++ x-pack/plugins/canvas/public/components/workpad_color_picker/index.ts
@@ -4,20 +4,5 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import { connect } from 'react-redux';
-import { addColor, removeColor } from '../../state/actions/workpad';
-import { getWorkpadColors } from '../../state/selectors/workpad';
-
-import { WorkpadColorPicker as Component } from '../workpad_color_picker/workpad_color_picker';
-import { State } from '../../../types';
-
-const mapStateToProps = (state: State) => ({
-  colors: getWorkpadColors(state),
-});
-
-const mapDispatchToProps = {
-  onAddColor: addColor,
-  onRemoveColor: removeColor,
-};
-
-export const WorkpadColorPicker = connect(mapStateToProps, mapDispatchToProps)(Component);
+export { WorkpadColorPicker } from './workpad_color_picker';
+export { WorkpadColorPicker as WorkpadColorPickerComponent } from './workpad_color_picker.component';
