+++ x-pack/plugins/canvas/public/components/workpad_color_picker/workpad_color_picker.component.tsx
