+++ x-pack/plugins/canvas/public/components/workpad_header/edit_menu/__examples__/edit_menu.stories.tsx
@@ -6,7 +6,7 @@
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import React from 'react';
-import { EditMenu } from '../edit_menu';
+import { EditMenu } from '../edit_menu.component';
 import { PositionedElement } from '../../../../../types';
 
 const handlers = {
