+++ x-pack/plugins/canvas/public/components/workpad_header/element_menu/__examples__/element_menu.stories.tsx
@@ -8,7 +8,7 @@ import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import React from 'react';
 import { ElementSpec } from '../../../../../types';
-import { ElementMenu } from '../element_menu';
+import { ElementMenu } from '../element_menu.component';
 
 const testElements: { [key: string]: ElementSpec } = {
   areaChart: {
