+++ x-pack/plugins/canvas/public/components/workpad_header/index.ts
@@ -4,5 +4,5 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export { EditMenu } from './edit_menu';
-export { EditMenu as EditMenuComponent } from './edit_menu.component';
+export { WorkpadHeader } from './workpad_header';
+export { WorkpadHeader as WorkpadHeaderComponent } from './workpad_header.component';
