+++ x-pack/plugins/canvas/public/components/workpad_header/refresh_control/refresh_control.component.tsx
