+++ x-pack/plugins/canvas/public/components/workpad_header/share_menu/__examples__/share_menu.stories.tsx
@@ -6,7 +6,7 @@
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import React from 'react';
-import { ShareMenu } from '../share_menu';
+import { ShareMenu } from '../share_menu.component';
 
 storiesOf('components/WorkpadHeader/ShareMenu', module).add('default', () => (
   <ShareMenu
