+++ x-pack/plugins/canvas/public/components/workpad_header/view_menu/view_menu.component.tsx
